+++ packages/plugins/typescript/graphql-request/tests/graphql-request.spec.ts
@@ -188,6 +188,36 @@ async function test() {
       expect(output).toMatchSnapshot();
     });
 
+    it('Should support emitLegacyCommonJSImports: false by emitting imports with extensions', async () => {
+      const config = { emitLegacyCommonJSImports: false };
+      const docs = [{ location: '', document: basicDoc }];
+      const result = (await plugin(schema, docs, config, {
+        outputFile: 'graphql.ts',
+      })) as Types.ComplexPluginOutput;
+
+      const usage = `
+async function test() {
+  const Client = require('graphql-request').GraphQLClient;
+  const client = new Client('');
+  const sdk = getSdk(client);
+
+  await sdk.feed();
+  await sdk.feed3();
+  await sdk.feed4();
+
+  const result = await sdk.feed2({ v: "1" });
+
+  if (result.feed) {
+    if (result.feed[0]) {
+      const id = result.feed[0].id
+    }
+  }
+}`;
+      const output = await validate(result, config, docs, schema, usage);
+
+      expect(output).toMatchSnapshot();
+    });
+
     it('Should support rawRequest when documentMode = "documentNode"', async () => {
       const config = { rawRequest: true };
       const docs = [{ location: '', document: basicDoc }];